Reorganized storage slots and variable types in BridgeState
All `*Timeout` values in `BridgeState` are represented as `uint32` now.
All `*RewardMultiplier` values in `BridgeState` are represented as `uint32` now.
`__depositAlignmentGap` changed to `bytes16` to really move `movingFundsTxMaxTotalFee`
to the next storage slot (previous `bytes8` version was not doing it).
Added `bytes16 __movingFundsAlignmentGap` to keep as one slot:
`uint96 movingFundsTimeoutSlashingAmount | uint32 movingFundsTimeoutNotifierRewardMultiplier | bytes16 movingFundsTimeoutSlashingGap`

This change allows to make timeout and reward multiplier values use consistent
types as well as reduce gas consumption a bit. Before vs after:

|  Bridge  ·  revealDeposit               ·  103961  ·  108365  ·  104924  │
|  Bridge  ·  revealDeposit               ·  103961  ·  108365  ·  104925  │

|  Bridge  ·  submitDepositSweepProof     ·  191029  ·  346180  ·  267960  │
|  Bridge  ·  submitDepositSweepProof     ·  191007  ·  346158  ·  267938  │

|  Bridge  ·  submitRedemptionProof       ·  126351  ·  197632  ·  168338  │
|  Bridge  ·  submitRedemptionProof       ·  126278  ·  197535  ·  168253  │

|  Bridge  ·  requestRedemption           ·  113740  ·  131575  ·  121303  │
|  Bridge  ·  requestRedemption           ·  113734  ·  131563  ·  121292  │

|  Bridge  ·  submitMovedFundsSweepProof  ·  134172  ·  147375  ·  139461  │
|  Bridge  ·  submitMovedFundsSweepProof  ·  134172  ·  147375  ·  139461  │

+++ solidity/contracts/bridge/Bridge.sol
@@ -190,9 +190,9 @@ contract Bridge is
         uint64 redemptionDustThreshold,
         uint64 redemptionTreasuryFeeDivisor,
         uint64 redemptionTxMaxFee,
-        uint64 redemptionTimeout,
+        uint32 redemptionTimeout,
         uint96 redemptionTimeoutSlashingAmount,
-        uint64 redemptionTimeoutNotifierRewardMultiplier
+        uint32 redemptionTimeoutNotifierRewardMultiplier
     );
 
     event MovingFundsParametersUpdated(
@@ -201,11 +201,11 @@ contract Bridge is
         uint32 movingFundsTimeoutResetDelay,
         uint32 movingFundsTimeout,
         uint96 movingFundsTimeoutSlashingAmount,
-        uint256 movingFundsTimeoutNotifierRewardMultiplier,
+        uint32 movingFundsTimeoutNotifierRewardMultiplier,
         uint64 movedFundsSweepTxMaxTotalFee,
         uint32 movedFundsSweepTimeout,
         uint96 movedFundsSweepTimeoutSlashingAmount,
-        uint64 movedFundsSweepTimeoutNotifierRewardMultiplier
+        uint32 movedFundsSweepTimeoutNotifierRewardMultiplier
     );
 
     event WalletParametersUpdated(
@@ -1228,9 +1228,9 @@ contract Bridge is
         uint64 redemptionDustThreshold,
         uint64 redemptionTreasuryFeeDivisor,
         uint64 redemptionTxMaxFee,
-        uint64 redemptionTimeout,
+        uint32 redemptionTimeout,
         uint96 redemptionTimeoutSlashingAmount,
-        uint64 redemptionTimeoutNotifierRewardMultiplier
+        uint32 redemptionTimeoutNotifierRewardMultiplier
     ) external onlyGovernance {
         self.updateRedemptionParameters(
             redemptionDustThreshold,
@@ -1311,11 +1311,11 @@ contract Bridge is
         uint32 movingFundsTimeoutResetDelay,
         uint32 movingFundsTimeout,
         uint96 movingFundsTimeoutSlashingAmount,
-        uint256 movingFundsTimeoutNotifierRewardMultiplier,
+        uint32 movingFundsTimeoutNotifierRewardMultiplier,
         uint64 movedFundsSweepTxMaxTotalFee,
         uint32 movedFundsSweepTimeout,
         uint96 movedFundsSweepTimeoutSlashingAmount,
-        uint64 movedFundsSweepTimeoutNotifierRewardMultiplier
+        uint32 movedFundsSweepTimeoutNotifierRewardMultiplier
     ) external onlyGovernance {
         self.updateMovingFundsParameters(
             movingFundsTxMaxTotalFee,
@@ -1607,9 +1607,9 @@ contract Bridge is
             uint64 redemptionDustThreshold,
             uint64 redemptionTreasuryFeeDivisor,
             uint64 redemptionTxMaxFee,
-            uint64 redemptionTimeout,
+            uint32 redemptionTimeout,
             uint96 redemptionTimeoutSlashingAmount,
-            uint64 redemptionTimeoutNotifierRewardMultiplier
+            uint32 redemptionTimeoutNotifierRewardMultiplier
         )
     {
         redemptionDustThreshold = self.redemptionDustThreshold;
@@ -1672,11 +1672,11 @@ contract Bridge is
             uint32 movingFundsTimeoutResetDelay,
             uint32 movingFundsTimeout,
             uint96 movingFundsTimeoutSlashingAmount,
-            uint256 movingFundsTimeoutNotifierRewardMultiplier,
+            uint32 movingFundsTimeoutNotifierRewardMultiplier,
             uint64 movedFundsSweepTxMaxTotalFee,
             uint32 movedFundsSweepTimeout,
             uint96 movedFundsSweepTimeoutSlashingAmount,
-            uint256 movedFundsSweepTimeoutNotifierRewardMultiplier
+            uint32 movedFundsSweepTimeoutNotifierRewardMultiplier
         )
     {
         movingFundsTxMaxTotalFee = self.movingFundsTxMaxTotalFee;
